fix: .bin data is uint32 not float32, fix intensity sum calculation
Becker & Hickl SPCImage .bin exports use 4-byte-per-element uint32
photon counts (0-8 per bin), not float32. Reading as float32 produced
denormalized near-zero values (~1e-45) that displayed as blank.

- Default dtype changed to uint32 in import dialog and auto-detect
- Intensity sum now handles both integer (int64 accumulator) and float
  dtypes correctly
- Verified: intensity range 0-24, 237K nonzero pixels, matches
  decay.sum(axis=-1) exactly

+++ src/percell4/gui/import_dialog.py
@@ -192,7 +192,7 @@ class ImportDialog(QDialog):
         bin_layout.addRow("Time bins:", self._bin_t)
 
         self._bin_dtype = QComboBox()
-        self._bin_dtype.addItems(["float32", "uint16", "uint32", "uint8"])
+        self._bin_dtype.addItems(["uint32", "uint16", "float32", "uint8"])
         bin_layout.addRow("Data type:", self._bin_dtype)
 
         self._bin_dim_order = QComboBox()
